jdk 21

+++ src/main/resources/db/initDB.sql
@@ -30,7 +30,7 @@ CREATE TABLE position
     user_id     INTEGER   NOT NULL,
     date_time   TIMESTAMP NOT NULL,
     symbol      TEXT      NOT NULL,
-    profit      DOUBLE PRECISION       NOT NULL,
+    profit      NUMERIC       NOT NULL,
     FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE
 );
 CREATE UNIQUE INDEX position_unique_user_datetime_idx ON position (user_id, date_time)
